requirements.txt
openai
python-dotenv
ollama
flask
gunicorn

